ppod buffer
