+++ code_examples/train_animalai/ppod/arena/hard_arena.yaml
@@ -1,0 +1,46 @@
+!ArenaConfig
+arenas:
+  0: !Arena
+    t: 2000
+    items:
+    - !Item
+      name: WallTransparent
+      positions:
+      - !Vector3 {x: 13.5, y: 0, z: 15}
+      sizes:
+      - !Vector3 {x: 4, y: 2, z: 4}
+      rotations: [90]
+    - !Item
+      name: WallTransparent
+      positions:
+      - !Vector3 {x: 22, y: 0, z: 15}
+      sizes:
+      - !Vector3 {x: 4, y: 2, z: 4}
+      rotations: [90]
+    - !Item
+      name: Cardbox1
+      positions:
+      - !Vector3 {x: 17.75, y: 0, z: 9}
+      sizes:
+      - !Vector3 {x: 4, y: 2, z: 4}
+      rotations: [90]
+    - !Item
+      name: Ramp
+      positions:
+      - !Vector3 {x: 28, y: 0, z: 15}
+      sizes:
+      - !Vector3 {x: 8, y: 2, z: 8}
+      colors:
+      - !RGB {r: 255, g: 0, b: 255}
+      rotations: [90]
+    - !Item
+      name: GoodGoal
+      positions:
+      - !Vector3 {x: 13.5, y: 2.25, z: 15}
+      sizes:
+      - !Vector3 {x: 3, y: 3, z: 3}
+    - !Item
+      name: Agent
+      positions:
+      - !Vector3 {x: 20, y: 0, z: 3}
+      rotations: [0]
