+++ code_examples/train_animalai/ppod/enjoy.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+DIR=code_examples/train_pybullet/ppod
+CUDA_VISIBLE_DEVICES="0" python $DIR/enjoy.py -c $DIR/conf.yaml
