+++ code_examples/train_animalai/ppod/plot.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+DIR=code_examples/train_pybullet/ppod
+python $DIR/plot.py -c $DIR/conf.yaml
