--- code_examples/train_sparse_reacher/ppod/conf.yaml
@@ -1,24 +1,0 @@
-num_env_steps : 10000000
-log_dir: /tmp/pybullet_ppo
-gamma: 0.99
-save_interval: 100000
-max_time: 86400
-num_env_processes: 12
-num_steps: 2048
-num_mini_batch: 32
-clip_param: 0.2
-ppo_epoch: 10
-gae_lambda: 0.95
-value_loss_coef: 0.5
-max_grad_norm: 0.5
-entropy_coef: 0.0
-lr: 0.0002
-nn: MLP
-frame_skip: 0
-frame_stack: 1
-env_id: HalfCheetahBulletEnv-v0
-num_grad_workers: 1 
-num_col_workers: 1 
-com_grad_workers: synchronous
-com_col_workers: synchronous
-
