--- code_examples/train_sparse_reacher/ppod/plot.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-DIR=code_examples/train_pybullet/ppod
-python $DIR/plot.py -c $DIR/conf.yaml
