--- code_examples/train_sparse_reacher/ppod/train.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-DIR=code_examples/train_sparse_reacher/ppod
-CUDA_VISIBLE_DEVICES="0" python $DIR/train.py -c $DIR/conf.yaml
